# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "Swagger Petstore",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "",
...
  "schemes": [
    "http"
  ],
  "security": [
    {
      "api_key": []
    }
  ],
  "paths": {
...
    },"/api/loginAPI": {
      "post": {
        "tags": [
          "loginAPI"
        ],
        "summary": "Add a new loginAPI",
        "description": "",
        "operationId": "loginAPI",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "loginAPI",
            "required": true,
            "schema": {
              "$ref": "#/definitions/loginAPI"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Valid"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "loginAPI": {
      "type": "object",
      "properties": {
        "nama_pengguna": {
          "type": "string",
          "example": "huft"
        },  
        "kata_sandi": {
          "type": "string",
          "example": "123"
        }
      }
    }
  },
...
}